+++ docs/main/changelog.rst
@@ -7,6 +7,11 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+Development version
+-------------------
+- Deprecate ``@containers.override()`` and ``@containers.copy()`` decorators.
+- Update changelog of version ``4.0.0`` so it lists all deprecated features.
+
 4.0.2
 -----
 - Fix typing stubs for ``@container.override()`` and ``@containers.copy()`` decorators (
@@ -33,6 +38,8 @@ Deprecations:
 - Deprecate ``ext.aiohttp`` module in favor of ``wiring`` feature.
 - Deprecate ``ext.flask`` module in favor of ``wiring`` feature.
 - Deprecate ``.delegate()`` provider method in favor of ``.provider`` attribute.
+- Deprecate ``@containers.override()`` decorator in favor of overriding container on instance level.
+- Deprecate ``@containers.copy()`` decorator.
 
 Removals: